Fix structure baseline regression after PR348 merge
Split newly extracted list operation and mobile menu modules into smaller focused files so the maintainability guardrail remains green without relaxing thresholds.
